+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Resolve layout shift on header focus (DR-3523)
 
 ### Added
-
+- log path in notFoundPage.tsx to new relic. (DR-3550)
 - Added Collections API authentication tokens (DR-3535)
 - Added search query redirects (DR-3529)
 - Connect search page to API (DR-3416)